feat(schema-bench): split LiamDB execution by dataset; add default/entity-extraction commands and shared runner
