+++ frontend/internal-packages/schema-bench/package.json
@@ -24,6 +24,8 @@
     "evaluateSchema": "tsx src/cli/evaluateSchema.ts",
     "evaluateSchemaMulti": "tsx src/cli/evaluateSchemaMulti.ts",
     "executeLiamDB": "tsx src/cli/executeLiamDb.ts",
+    "executeLiamDB:default": "tsx src/cli/executeLiamDbDefault.ts",
+    "executeLiamDB:entityExtraction": "tsx src/cli/executeLiamDbEntityExtraction.ts",
     "executeLiamDBMulti": "tsx src/cli/executeLiamDbMulti.ts",
     "executeOpenai": "tsx src/cli/executeOpenai.ts",
     "fmt": "concurrently \"pnpm:fmt:*\"",
